Update routing tests
